Refactoring of SalesByDayOfWeekData
